requirements.txt
tensorflow-cpu~=2.5.3
pandas~=1.2.3
numpy~=1.19.5
biopython~=1.81